Move HorizontalVelocityChangeEvent to event package
